Move config structs to pkg/config
Need to use these structs in the CRD and controller
for container runtime config. Moving the structs to
a separate package to reduce dependencies.
