+++ server/config.go
@@ -6,6 +6,7 @@ import (
 
 	"github.com/BurntSushi/toml"
 	"github.com/kubernetes-sigs/cri-o/lib"
+	"github.com/kubernetes-sigs/cri-o/pkg/config"
 )
 
 const (
@@ -17,40 +18,7 @@ const (
 // the server. This is intended to be loaded from a toml-encoded config file.
 type Config struct {
 	lib.Config
-	APIConfig
-}
-
-// APIConfig represents the "crio.api" TOML config table.
-type APIConfig struct {
-	// GRPCMaxSendMsgSize is the maximum grpc send message size in bytes.
-	GRPCMaxSendMsgSize int `toml:"grpc_max_send_msg_size"`
-
-	// GRPCMaxRecvMsgSize is the maximum grpc receive message size in bytes.
-	GRPCMaxRecvMsgSize int `toml:"grpc_max_recv_msg_size"`
-
-	// Listen is the path to the AF_LOCAL socket on which cri-o will listen.
-	// This may support proto://addr formats later, but currently this is just
-	// a path.
-	Listen string `toml:"listen"`
-
-	// StreamAddress is the IP address on which the stream server will listen.
-	StreamAddress string `toml:"stream_address"`
-
-	// StreamPort is the port on which the stream server will listen.
-	StreamPort string `toml:"stream_port"`
-
-	// StreamEnableTLS enables encrypted tls transport of the stream server
-	StreamEnableTLS bool `toml:"stream_enable_tls"`
-
-	// StreamTLSCert is the x509 certificate file path used to serve the encrypted stream
-	StreamTLSCert string `toml:"stream_tls_cert"`
-
-	// StreamTLSKey is the key file path used to serve the encrypted stream
-	StreamTLSKey string `toml:"stream_tls_key"`
-
-	// StreamTLSCA is the x509 CA(s) file used to verify and authenticate client
-	// communication with the tls encrypted stream
-	StreamTLSCA string `toml:"stream_tls_ca"`
+	config.APIConfig
 }
 
 // tomlConfig is another way of looking at a Config, which is
@@ -58,11 +26,11 @@ type APIConfig struct {
 // conversions.
 type tomlConfig struct {
 	Crio struct {
-		lib.RootConfig
-		API     struct{ APIConfig }         `toml:"api"`
-		Runtime struct{ lib.RuntimeConfig } `toml:"runtime"`
-		Image   struct{ lib.ImageConfig }   `toml:"image"`
-		Network struct{ lib.NetworkConfig } `toml:"network"`
+		config.RootConfig
+		API     struct{ config.APIConfig }     `toml:"api"`
+		Runtime struct{ config.RuntimeConfig } `toml:"runtime"`
+		Image   struct{ config.ImageConfig }   `toml:"image"`
+		Network struct{ config.NetworkConfig } `toml:"network"`
 	} `toml:"crio"`
 }
 
@@ -124,7 +92,7 @@ func (c *Config) ToFile(path string) error {
 func DefaultConfig() *Config {
 	return &Config{
 		Config: *lib.DefaultConfig(),
-		APIConfig: APIConfig{
+		APIConfig: config.APIConfig{
 			Listen:             CrioSocketPath,
 			StreamAddress:      "127.0.0.1",
 			StreamPort:         "0",
